requirements.txt
flask==3.1.1
flask-cors==6.0.1
pillow>=11.0.0
pytest
